implement test for new event using env.event
